requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
ccxt==4.3.88
pydantic==2.8.2
python-dotenv==1.0.1
loguru==0.7.2
SQLAlchemy==2.0.31
Jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
cryptography==42.0.8
requests==2.32.3
starlette==0.37.2
